model as function plus pandas database integration
I added the model as a function. This makes it easier to run multiple
times. The model now keeps track of all variables in pandas dataframes
instead writing to the SQL database on the go. I do believe that this
made the model much slower. So a next step would be to speed it up
again.